Refactored date comparator SqlFunctions to become ProgramIndicatorFunctions
