+++ dhis-2/dhis-services/dhis-service-core/src/main/java/org/hisp/dhis/program/MinutesBetweenProgramIndicatorFunction.java
@@ -1,4 +1,4 @@
-package org.hisp.dhis.commons.sqlfunc;
+package org.hisp.dhis.program;
 
 /*
  * Copyright (c) 2004-2018, University of Oslo
@@ -29,18 +29,18 @@ package org.hisp.dhis.commons.sqlfunc;
  */
 
 /**
- * Function which evaluates to the number of months between two given dates.
+ * Function which evaluates to the number of minutes between two given dates.
  *
  * @author Markus Bekken
  */
-public class MonthsBetweenSqlFunction
-    extends BaseDateComparatorSqlFunction
+public class MinutesBetweenProgramIndicatorFunction
+    extends BaseDateComparatorProgramIndicatorFunction
 {
-    public static final String KEY = "monthsBetween";
+    public static final String KEY = "minutesBetween";
 
     @Override
     protected String compare( String startDate, String endDate )
     {
-        return "(date_part('month',age(cast(" + endDate + " as date), cast(" + startDate + " as date))))";
+        return "(extract(epoch from (cast(" + endDate + " as timestamp) - cast(" + startDate + " as timestamp))) / 60)";
     }
 }
