+++ dhis-2/dhis-services/dhis-service-core/src/main/java/org/hisp/dhis/program/WeeksBetweenProgramIndicatorFunction.java
@@ -1,4 +1,4 @@
-package org.hisp.dhis.commons.sqlfunc;
+package org.hisp.dhis.program;
 
 /*
  * Copyright (c) 2004-2018, University of Oslo
@@ -29,18 +29,18 @@ package org.hisp.dhis.commons.sqlfunc;
  */
 
 /**
- * Function which evaluates to the number of minutes between two given dates.
+ * Function which evaluates to the number of weeks between two given dates.
  *
  * @author Markus Bekken
  */
-public class MinutesBetweenSqlFunction
-    extends BaseDateComparatorSqlFunction
+public class WeeksBetweenProgramIndicatorFunction
+    extends BaseDateComparatorProgramIndicatorFunction
 {
-    public static final String KEY = "minutesBetween";
+    public static final String KEY = "weeksBetween";
 
     @Override
     protected String compare( String startDate, String endDate )
     {
-        return "(extract(epoch from (cast(" + endDate + " as timestamp) - cast(" + startDate + " as timestamp))) / 60)";
+        return "((cast(" + endDate + " as date) - cast(" + startDate + " as date))/7)";
     }
 }
