+++ dhis-2/dhis-services/dhis-service-core/src/main/java/org/hisp/dhis/program/YearsBetweenProgramIndicatorFunction.java
@@ -1,4 +1,4 @@
-package org.hisp.dhis.commons.sqlfunc;
+package org.hisp.dhis.program;
 
 /*
  * Copyright (c) 2004-2018, University of Oslo
@@ -29,31 +29,18 @@ package org.hisp.dhis.commons.sqlfunc;
  */
 
 /**
- * Function which evaluates a relation between two given dates.
+ * Function which evaluates to the number of years between two given dates.
  *
  * @author Markus Bekken
  */
-public abstract class BaseDateComparatorSqlFunction
-    implements SqlFunction
+public class YearsBetweenProgramIndicatorFunction
+    extends BaseDateComparatorProgramIndicatorFunction
 {
-    protected abstract String compare( String startDate, String endDate );
+    public static final String KEY = "yearsBetween";
 
     @Override
-    public String evaluate( String... args )
+    protected String compare( String startDate, String endDate )
     {
-        if ( args == null || args.length != 2 )
-        {
-            throw new IllegalArgumentException( "Illegal arguments, expected 2 arguments: start-date, end-date" );
-        }
-
-        String startDate = args[0];
-        String endDate = args[1];
-
-        return compare( startDate, endDate );
-    }
-
-    public String getSampleValue()
-    {
-        return "1";
+        return "(date_part('year',age(cast(" + endDate + " as date), cast(" + startDate + " as date))))";
     }
 }
